dodanie obsługi wyjątków w SWIG-owych wrapperach, dodanie wrappera dla Perla
git-svn-id: svn://svn.nlp.ipipan.waw.pl/morfeusz/morfeusz@184 ff4e3ee1-f430-4e82-ade0-24591c43f1fd

+++ morfeusz/Qualifiers.cpp
@@ -36,8 +36,8 @@ vector<string> Qualifiers::getQualifiers(int n) const {
     return this->qualifiers.at(n);
 }
 
-int Qualifiers::getQualifiersNum() const {
-    return this->qualifiers.size();
+unsigned int Qualifiers::getQualifiersNum() const {
+    return (unsigned int) this->qualifiers.size();
 }
 
 Qualifiers::~Qualifiers() {
